Update website via upload
All menus now have their items on the right hand side with a simple "TonySax.org" label on the left. All pages also have a updated footer. On the frontpage, a caption has been added to the picture. The Projects page now has a title and a panel that lists my repositories. The Photos page now has a "pagination" and several tiles that display randomly picked photos.