Updated all files plus added bulma.css file
Revamped the home page. It is now called home.html instead of frontpage.html. Made minor changes to about.html. Put a modal on travel.html. Added progress bars to projects.html and other minor changes. Fixed the spacing of the gallery on photos16x9.html. Added a title bar and picture as links on the contact page. The footer was also slightly changed.