Latest Bulma version for all files. Hamburger bar.
Switched to the latest version of Bulma for the files that were using the old version. Added the functionality for the hamburger bar for mobile and tablet without using javascript. Fixed spacing on the title page. Changed the display for mobile on the about page. Removed some spacing on the about page.